Update delete subcommand to ask for confirmation

+++ src/main.rs
@@ -394,8 +394,27 @@ fn main() -> Result<()> {
             let working_date = WorkDate::now();
 
             let task_id = db.get_task_id_by_seqnum(opts.task_number, working_date)?;
-            db.delete_task(task_id)?;
-            println!("task {} deleted", opts.task_number);
+            let task = db.get_task(task_id)?;
+
+            eprintln!(
+                "\"{}\" started at {} {}",
+                task.name(),
+                task.working_date().to_string(),
+                task.start_time().to_string_hhmm()
+            );
+
+            let proceed = Confirm::new()
+                .with_prompt("Really delete?")
+                .wait_for_newline(false)
+                .default(false)
+                .show_default(true)
+                .interact()?;
+            if proceed {
+                db.delete_task(task_id)?;
+                println!("\ntask {} deleted", opts.task_number);
+            } else {
+                println!("\nOparation canceled.");
+            };
         }
 
         SubCommand::ShowManager => {
